% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_files_to_one_df.R
\name{compile_distance_to_one_df}
\alias{compile_distance_to_one_df}
\title{Compile distance to one data frame}
\usage{
compile_distance_to_one_df(b, v, start_times, path = NULL)
}
\arguments{
\item{b}{required, 8 digit building number, used in name the output file}

\item{v}{required, variable to compile}

\item{start_times}{required, a vector of start time for each individual files}

\item{path}{optional, directory to save downloaded data}
}
\description{
This function concatenates year by year downloaded distance_files into one data frame
}
\examples{
compile_weather_to_one_df (b="xxxxxxxx", v, start_times, path=NULL)
}
\keyword{concatenate}
