% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_noaa_ghcnd.R
\name{compile_weather_ghcnd_main}
\alias{compile_weather_ghcnd_main}
\title{Compile weather ghcnd main routine}
\usage{
compile_weather_ghcnd_main(building, station_df, date_min = NULL,
  date_max = NULL, var = "TMIN",
  format_fun = get.noaa.weather::degreeCtoF)
}
\arguments{
\item{building}{required, will be a column of returned data frame}

\item{station_df}{required, a data frame containing the ids of nearby stations}

\item{var}{required, the variable to compile, default to "TMIN"}
}
\description{
This function downloads and compiles ghcnd files
}
\examples{
for one building: compile_weather_isd_main(useSavedData=FALSE, years=c(2015, 2016, 2017), building="AK0000ZZ")
}
\keyword{download}
\keyword{file}
\keyword{isd}
