% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_noaa_isd.R
\name{compile_weather_isd_main}
\alias{compile_weather_isd_main}
\title{Compile weather isd main routine}
\usage{
compile_weather_isd_main(useSavedData, years, lat_lon_df, latitude, longitude,
  building)
}
\arguments{
\item{useSavedData}{required, if you want to use saved list of
station_to_download_year.feather in the saved directory}

\item{years}{required, a vector of years to compile, e.g. years = c(2015,
2016, 2017)}

\item{latitude}{optional, latitude of building, if latitude is supplied, we
assume longitude and building are also supplied}

\item{longitude}{optional, longitude of building}

\item{building}{optional, if supplied, only compile results for a building,
if missing, if missing, all building's latitude and longitude will be
used}

\item{var}{required, variable to compile}
}
\description{
This function downloads and compiles noaa isd files rnoaa
rnoaa::isd_stations(refresh = TRUE), but will be a bit slow
}
\examples{
for all buildings: compile_weather_isd_main(useSavedData=TRUE, years=c(2015, 2016, 2017))
for one building: compile_weather_isd_main(useSavedData=FALSE, years=c(2015, 2016, 2017), building="AK0000ZZ")
}
\keyword{download}
\keyword{file}
\keyword{isd}
