% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_noaa_isd.R
\name{degreeCtoF}
\alias{degreeCtoF}
\title{Unit conversion from degree C to F}
\usage{
degreeCtoF(c)
}
\arguments{
\item{c}{required, degree c to convert}
}
\description{
Unit conversion from degree C to F
}
\examples{
degreeCtoF(30)
}
\keyword{conversion}
\keyword{unit}
