% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_noaa_isd.R
\name{download_isd}
\alias{download_isd}
\title{Download isd files to cache directory, with error handler for download failure}
\usage{
download_isd(station_to_download, year, start, end)
}
\arguments{
\item{station_to_download}{required, a data frame with two columns "usaf",
and "wban"}

\item{year}{required, the year of weather data to download}

\item{start}{optional, the start index of the weather stations to download.
This allows users to download part of the stations in station_to_download}

\item{end}{optional, the end index of the weather stations to download. This
allows users to download part of the stations in station_to_download}
}
\description{
This function downloads unique stations in station_to_download to a cache
directory in rnoaa, it allows user to specify the starting and ending index
to allow for downloading part of the stations in station_to_download
}
\examples{
download_isd(station_to_download, year=year, variables=variables)
}
\keyword{download}
\keyword{file}
\keyword{isd}
