% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_noaa_ghcnd.R
\name{download_stations}
\alias{download_stations}
\title{Download a list of stations, return the bad ones}
\usage{
download_stations(stations, date_min, date_max, v, v_out)
}
\arguments{
\item{stations}{the list of stations to download, normally much larger than limit}

\item{date_min}{a string "yyyy-mm-dd"}

\item{date_max}{a string "yyyy-mm-dd"}

\item{v}{required, variable to download, e.g. TMIN}

\item{v_out}{required, the output variable from noaa ghcnd, e.g. tmin}
}
\description{
Download a list of stations, return the bad ones
}
\examples{
get_nearby_isd_stations(lat_lon_df, isd_data=isd_data, radius = 100, limit=5,
  date_min = 20150101, date_max = 20151231)
}
\keyword{isd}
\keyword{nearby}
