% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_noaa_isd.R
\name{get_isdfile_by_year}
\alias{get_isdfile_by_year}
\title{Download isd files to cache directory, and record station mappings to feather files}
\usage{
get_isdfile_by_year(year, lat_lon_df, isd_data)
}
\arguments{
\item{year}{required, the year of isd weather data to download}

\item{lat_lon_df}{optional, default to use db.interface::get_lat_lon_df}

\item{isd_data}{optional, if missing, it will be created with
rnoaa::isd_stations(refresh = TRUE), but will be a bit slow}
}
\description{
This function downloads unique stations in lat_lon_df to a cache directory in
rnoaa
}
\examples{
get_isdfile_by_year(year)
}
\keyword{download}
\keyword{file}
\keyword{isd}
