% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_noaa_ghcnd.R
\name{get_nearby_ghcnd_stations_one_loc}
\alias{get_nearby_ghcnd_stations_one_loc}
\title{Get a data frame of nearby weather stations for one location}
\usage{
get_nearby_ghcnd_stations_one_loc(lat_lon_df, id_col_name = "id", ghcnd_data,
  v, radius = NULL, limit = NULL, date_min = NULL, date_max = NULL,
  year = NULL)
}
\arguments{
\item{lat_lon_df}{a data frame with one row, containing a column "latitude",
a column "longitude", and an id column}

\item{id_col_name}{optional, a unique identifier for each location, default
to be "id"}

\item{ghcnd_data}{optional, returned by rnoaa::ghcnd_stations(refresh = TRUE)}

\item{v}{required, variable to download}

\item{radius}{(numeric) Radius (in km) to search from the lat,lon
coordinates, used in isd_station_search}

\item{limit}{the maximum nearest stations returned for each building}

\item{date_min}{An integer giving the earliest date of the weather time
series that the user would like in the final output. This integer should be
formatted as yyyymmdd (20150101 for Jan 1, 2015)}

\item{date_max}{An integer giving the latest date of the weather time series
that the user would like in the final output. This integer should be
formatted as yyyymmdd (20150101 for Jan 1, 2015)}

\item{v}{optional, variable to download, default to TMIN}

\item{testing}{optional, if marked true, only the head part of lat_lon_df
will be evaluated}
}
\description{
This function get a data frame of nearby stations for one location, when
there is missing data, we want to exclude the station and re-download, until
no missing data
}
\examples{
get_nearby_isd_stations(lat_lon_df, isd_data=isd_data, radius = 100, limit=5,
  date_min = 20150101, date_max = 20151231)
}
\keyword{isd}
\keyword{nearby}
