% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_noaa_isd.R
\name{get_nearby_isd_stations}
\alias{get_nearby_isd_stations}
\title{Get a data frame of nearby weather stations}
\usage{
get_nearby_isd_stations(lat_lon_df, isd_data, radius, limit, date_min, date_max,
  year)
}
\arguments{
\item{lat_lon_df}{a data frame containing a column "latitude", and a column
"longitude"}

\item{isd_data}{returned by rnoaa::isd_stations(refresh = TRUE)}

\item{radius}{(numeric) Radius (in km) to search from the lat,lon
coordinates, used in isd_station_search}

\item{limit}{the maximum nearest stations returned for each building}

\item{date_min}{An integer giving the earliest date of the weather time
series that the user would like in the final output. This integer should be
formatted as yyyymmdd (20150101 for Jan 1, 2015)}

\item{date_max}{An integer giving the latest date of the weather time series
that the user would like in the final output. This integer should be
formatted as yyyymmdd (20150101 for Jan 1, 2015)}
}
\description{
This function get a data frame of nearby stations, with columns: usaf, wban,
begin, end, distance, and Building_Number
}
\examples{
get_nearby_isd_stations(lat_lon_df, isd_data=isd_data, radius = 100, limit=5,
  date_min = 20150101, date_max = 20151231)
}
\keyword{isd}
\keyword{nearby}
