% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_noaa_isd.R
\name{get_unique_stations}
\alias{get_unique_stations}
\title{Get unique station ids in two columns: usaf, and wban}
\usage{
get_unique_stations(station_df)
}
\arguments{
\item{station_df}{a data frame containing at least the two columns: usaf, and wban}
}
\description{
This function get a data frame of unique isd stations, with columns: usaf, wban,
}
\examples{
lat_lon_df = db.interface::get_lat_lon_df()
get_unique_stations(lat_lon_df)
}
\keyword{isd}
\keyword{nearby}
