% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_noaa_ghcnd.R
\name{nameListToDf}
\alias{nameListToDf}
\title{Convert named list to a data frame}
\usage{
nameListToDf(nameList, nameCol)
}
\arguments{
\item{nameList}{the input named list, each element is a data frame}

\item{nameCol}{the column name to put the names in of the named list}
}
\description{
This function converts a named list data frame to a data frame, with a column
nameCol tagging the names of the named list
}
\examples{
nameListToDf(nearbyStations, "building")
}
\keyword{df}
\keyword{list}
\keyword{named}
\keyword{to}
