% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_noaa_isd.R
\name{read_var_by_year}
\alias{read_var_by_year}
\title{Compute weighted average temperature of nearby stations for buildings in station_df}
\usage{
read_var_by_year(station_df, var, format_fun, year)
}
\arguments{
\item{station_df}{required, a dataframe containing station id (usaf and wban), and Building_Number}

\item{var}{required, the variable to compile, you can choose from the 4:
"temperature", "temperature_dewpoint", "wind_direction", "wind_speed"}

\item{format_fun}{required, function to format variable}

\item{year}{required, year of isd file to compile}
}
\description{
This function downloads and compiles noaa isd files rnoaa
rnoaa::isd_stations(refresh = TRUE), but will be a bit slow
}
\examples{
read_var_by_year(station_df, var="temperature", format_fun=format_noaa_temperature, year=2015)
}
\keyword{download}
\keyword{file}
\keyword{isd}
