% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_download.R
\name{run_download}
\alias{run_download}
\title{Run download of weather data}
\usage{
run_download(download_rounds, start_times, end_times, v, ghcnd_data_var,
  radius = 100, path = NULL)
}
\arguments{
\item{path}{optional, directory to save downloaded data.}

\item{b}{required, 8 digit building number}

\item{start_str}{time to start downloading, a date string of yyyy-mm-dd format}

\item{end_str}{time to end downloading, a date string of yyyy-mm-dd format}
}
\description{
This function downloads data from noaa ghcnd
computes a inverse weighted average of the temperature
}
\examples{
getAvgTemp(building="UT0032ZZ", start_str="2012-01-05", end_str="2015-01-30")
}
\keyword{average}
\keyword{daily}
\keyword{noaa}
\keyword{temperature}
