% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapViews.R
\name{view_isd_stations_year}
\alias{view_isd_stations_year}
\title{plot isd_stations}
\usage{
view_isd_stations_year(isd_data, year)
}
\arguments{
\item{isd_data}{optional, returned by rnoaa::isd_stations(refresh = TRUE)}

\item{year}{optional, if supplied only plot stations with data in that year}
}
\description{
This function plots the isd stations available for a year
}
\examples{
view_isd_stations_year(isd_data, year=2015)
}
\keyword{isd}
\keyword{plot}
\keyword{station}
